feat: Adds a Gallery Browser
This commit adds a new Gallery Browser to the app.

It allows the user to view the pictures in a grid, loaded from the data source. (Currently using a test api randomuser.me) but in future this is supposed to be the camera roll of the user or maybe a cloud api like google drive/photos/dropbox etc.

As part of the first iteration this commit only shows a sample drawable in the grid.